+++ src/charm_state.py
@@ -36,6 +36,7 @@ ARCHITECTURES_X86 = {"x86_64"}
 
 CHARM_STATE_PATH = Path("charm_state.json")
 
+ALLOW_EXTERNAL_CONTRIBUTOR_CONFIG_NAME = "allow-external-contributor"
 BASE_VIRTUAL_MACHINES_CONFIG_NAME = "base-virtual-machines"
 DOCKERHUB_MIRROR_CONFIG_NAME = "dockerhub-mirror"
 FLAVOR_LABEL_COMBINATIONS_CONFIG_NAME = "flavor-label-combinations"
@@ -200,6 +201,13 @@ class RepoPolicyComplianceConfig(BaseModel):
                 f"Missing {REPO_POLICY_COMPLIANCE_URL_CONFIG_NAME} configuration"
             )
 
+        logger.warning(
+            "%s, %s configuration option is marked for deprecation."
+            "Consider using %s configuration option instead.",
+            REPO_POLICY_COMPLIANCE_URL_CONFIG_NAME,
+            REPO_POLICY_COMPLIANCE_TOKEN_CONFIG_NAME,
+            ALLOW_EXTERNAL_CONTRIBUTOR_CONFIG_NAME,
+        )
         # pydantic allows string to be passed as AnyHttpUrl, mypy complains about it
         return cls(url=url, token=token)  # type: ignore
 
@@ -210,6 +218,9 @@ class CharmConfig(BaseModel):
     Some charm configurations are grouped into other configuration models.
 
     Attributes:
+        allow_external_contributor: Whether to allow runs from forked repositories with from
+            an external contributor with author association status less than COLLABORATOR. See \
+            https://docs.github.com/en/graphql/reference/enums#commentauthorassociation.
         dockerhub_mirror: Private docker registry as dockerhub mirror for the runners to use.
         labels: Additional runner labels to append to default (i.e. os, flavor, architecture).
         openstack_clouds_yaml: The openstack clouds.yaml configuration.
@@ -226,6 +237,7 @@ class CharmConfig(BaseModel):
         runner_manager_log_level: The log level of the runner manager application.
     """
 
+    allow_external_contributor: bool
     dockerhub_mirror: AnyHttpsUrl | None
     labels: tuple[str, ...]
     openstack_clouds_yaml: OpenStackCloudsYAML
@@ -501,7 +513,7 @@ class CharmConfig(BaseModel):
         manager_proxy_command = (
             cast(str, charm.config.get(MANAGER_SSH_PROXY_COMMAND_CONFIG_NAME, "")) or None
         )
-        use_aproxy = bool(charm.config.get(USE_APROXY_CONFIG_NAME))
+        use_aproxy = cast(bool, charm.config.get(USE_APROXY_CONFIG_NAME, False))
 
         custom_pre_job_script = (
             cast(str, charm.config.get(CUSTOM_PRE_JOB_SCRIPT_CONFIG_NAME, "")) or None
@@ -512,6 +524,9 @@ class CharmConfig(BaseModel):
         )
         # pydantic allows to pass str as AnyHttpUrl, mypy complains about it
         return cls(
+            allow_external_contributor=cast(
+                bool, charm.config.get(ALLOW_EXTERNAL_CONTRIBUTOR_CONFIG_NAME, False)
+            ),
             dockerhub_mirror=dockerhub_mirror,  # type: ignore
             labels=labels,
             openstack_clouds_yaml=openstack_clouds_yaml,