+++ docs/changelog.md
@@ -6,6 +6,11 @@ This changelog documents user-relevant changes to the GitHub runner charm.
 
 - Added timeouts to API calls (Openstack,GitHub,Repo policy compliance) to fix a hanging GitHub runner manager application. 
 
+## 2025-11-30
+
+- Added a new configuration option `allow-external-contributor` as an non-external replacement for
+  `repo-policy-compliance` service.
+
 ## 2025-11-21
 
 - Updated broken links in `docs/how-to/index.md`.